requirements.txt
asyncio
aiogram<4.0
python-dotenv
fake_useragent